Organise includes for util and benchmarking
