+++ include/sddf/util/util.h
