+++ network/components/lwip_timer.c
@@ -1,4 +1,4 @@
-#include "timer.h"
+#include <sddf/util/timer.h>
 
 #define TIMER_CH 1 
 #define LWIP_TICK_MS 100
