+++ network/components/udp_echo_socket.c
@@ -16,8 +16,8 @@
 #include "lwip/pbuf.h"
 #include "lwip/udp.h"
 
-#include "echo.h"
-#include "util.h"
+#include <sddf/util/echo.h>
+#include <sddf/util/util.h>
 
 #define NUM_LOOPS 10
 #define UDP_ECHO_PORT 1235
