+++ network/ipstacks/lwip/src/include/lwip/arch.h
@@ -45,7 +45,7 @@
 #define BIG_ENDIAN 4321
 #endif
 
-#include "arch/cc.h"
+#include <sddf/util/arch/cc.h>
 
 /**
  * @defgroup compiler_abstraction Compiler/platform abstraction
