+++ network/ipstacks/lwip/src/include/lwip/opt.h
@@ -48,7 +48,7 @@
  * Include user defined options first. Anything not defined in these files
  * will be set to standard values. Override anything you don't like!
  */
-#include "lwipopts.h"
+#include <sddf/util/lwipopts.h>
 #include "lwip/debug.h"
 
 /**
